Organized emitters better
