+++ src/main/java/personal/opensrcerer/reactive/emitters/components/ButtonClickEmitter.java
@@ -1,4 +1,4 @@
-package personal.opensrcerer.reactive.emitters.emitters;
+package personal.opensrcerer.reactive.emitters.components;
 
 import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
 import personal.opensrcerer.reactive.emitters.DiscordEventEmitter;
