+++ src/main/java/personal/opensrcerer/reactive/emitters/system/ReadyEmitter.java
@@ -1,4 +1,4 @@
-package personal.opensrcerer.reactive.emitters.emitters;
+package personal.opensrcerer.reactive.emitters.system;
 
 import net.dv8tion.jda.api.events.ReadyEvent;
 import personal.opensrcerer.reactive.sinks.system.ReadySink;
