--- src/main/java/personal/opensrcerer/sinks/slash/SlashCommandSink.java
@@ -1,7 +1,0 @@
-package personal.opensrcerer.sinks.slash;
-
-import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
-import personal.opensrcerer.sinks.Sink;
-
-public abstract class SlashCommandSink implements Sink<SlashCommandEvent> {
-}
